remove landing garbage, fix docs and account menu colors, fix landing left menu
