--- src/dotnet/UI.Blazor.App/Pages/Landing/FeaturesCarousel.razor
@@ -1,153 +1,0 @@
-@{
-    var title = ScreenSize.IsNarrow() ? "Actual Chat Features" : "Features";
-}
-
-<LandingCarousel
-    CarouselOrder="2"
-    CarouselLength="6"
-    Class="feature-carousel"
-    WrapperClass="pt-4 px-4 text-headline-4 md:text-5.5xl"
-    WrapperTitle="@title">
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="1">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-1.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-1.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="font-bold">
-                Recent chats
-            </div>
-            <div>
-                chats that we have communicated in recently and where we can go back to by click
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="2">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-2.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-2.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="font-bold">
-                Recent chats
-            </div>
-            <div>
-                chats that we have communicated in recently and where we can go back to by click
-            </div>
-            <div class="mt-4 font-bold">
-                Group and peer-to-peer chats
-            </div>
-            <div class="mt-4 font-bold">
-                Public <span class="font-normal">(anyone can view & listen)</span> and private chats
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="3">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-3.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-3.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="font-bold">
-                Voice & text
-            </div>
-            <div>
-                are blended together– everywhere!
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="4">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-4.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-4.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="font-bold">
-                Voice & text
-            </div>
-            <div>
-                are blended together– everywhere!
-            </div>
-            <div class="font-bold">
-                Replies, @@mentions, and reactions
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="5">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-5.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-5.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="font-bold">
-                Voice & text
-            </div>
-            <div>
-                are blended together– everywhere!
-            </div>
-            <div class="mt-4 font-bold">
-                Replies, @@mentions, and reactions
-            </div>
-            <div class="mt-4 font-bold">
-                Playing messages
-            </div>
-            <div>
-                Read and listen to received voice messages at any time
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-
-    <FeaturesCarouselItem
-        Class="carousel-page bg-[#EEEEF6]"
-        SlideOrder="6">
-        <ImageContent>
-            @if (ScreenSize.IsNarrow()) {
-                <img draggable="false" src="/dist/images/landing/features-carousel-mobile-6.png" alt="">
-            } else {
-                <img draggable="false" src="/dist/images/landing/features-carousel-6.png" alt="">
-            }
-        </ImageContent>
-        <Description>
-            <div class="mt-4 font-bold">
-                Avatars
-            </div>
-            <div>
-                change name, picture, and bio on per-chat basis
-            </div>
-        </Description>
-    </FeaturesCarouselItem>
-</LandingCarousel>
-
-@code {
-    [CascadingParameter] ScreenSize ScreenSize { get; set; }
-}
