--- src/dotnet/UI.Blazor.App/Pages/Landing/FeaturesCarouselItem.razor
@@ -1,30 +1,0 @@
-<div class="page full-screen-page @Class" id="@GetCurrentSlideId()">
-    <div class="features-carousel-item">
-        <div class="c-title">
-            @Title
-        </div>
-        <div class="c-image">
-            @ImageContent
-        </div>
-        <div class="c-description">
-            @Description
-        </div>
-    </div>
-</div>
-
-@code {
-    [CascadingParameter(Name = "CarouselLength")] public int CarouselLength { get; set; }
-    [CascadingParameter(Name = "CarouselOrder")] public int CarouselOrder { get; set; }
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public RenderFragment? Title { get; set; }
-    [Parameter] public RenderFragment? ImageContent { get; set; }
-    [Parameter] public RenderFragment? Description { get; set; }
-    [Parameter] public int SlideOrder { get; set; }
-
-    private string GetCurrentSlideId() {
-        if (CarouselOrder != 0 && SlideOrder != 0) {
-            return $"slide-{CarouselOrder}-{SlideOrder}";
-        }
-        return "";
-    }
-}
