--- src/dotnet/UI.Blazor.App/Pages/Landing/ImagePosition.cs
@@ -1,9 +1,0 @@
-namespace ActualChat.UI.Blazor.App.Pages.Landing;
-
-public enum ImagePosition
-{
-    Top,
-    Right,
-    Bottom,
-    Left,
-}
