+++ src/dotnet/UI.Blazor.App/Pages/Landing/Landing.razor
@@ -7,9 +7,9 @@
 @layout NoLayout
 
 @if (HostInfo.HostKind.IsMauiApp()) {
-    <NewLandingApp/>
+    <LandingApp/>
 } else {
-    <NewLandingWeb/>
+    <LandingWeb/>
 }
 
 @code {
