--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingCarousel.razor
@@ -1,25 +1,0 @@
-<div class="carousel scrollable @Class" id="carousel-@CarouselOrder">
-    <CascadingValue Name="CarouselLength" Value="CarouselLength">
-        <CascadingValue Name="CarouselOrder" Value="CarouselOrder">
-            <LandingCarouselWrapper>
-                <TopContent>
-                    <div class="@WrapperClass">
-                        @WrapperTitle
-                    </div>
-                </TopContent>
-            </LandingCarouselWrapper>
-            <div class="carousel-content">
-                @ChildContent
-            </div>
-        </CascadingValue>
-    </CascadingValue>
-</div>
-
-@code {
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-    [Parameter] public int CarouselOrder { get; set; }
-    [Parameter] public int CarouselLength { get; set; }
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public string WrapperTitle { get; set; } = "";
-    [Parameter] public string WrapperClass { get; set; } = "";
-}
