--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingCarouselWrapper.razor
@@ -1,43 +1,0 @@
-@{
-    var currentSlideId = GetCurrentSlideId();
-}
-
-<div class="carousel-wrapper @Class" name="@currentSlideId">
-    <div class="c-side-left">
-        <div class="arrow">
-            <i class="icon-chevron-left text-5xl text-[#9294E1]"></i>
-        </div>
-    </div>
-    <div class="c-body">
-        <div class="c-top">
-            @TopContent
-        </div>
-        <div class="c-bottom">
-            @for (var i = 1; i <= CarouselLength; i++) {
-                var slideId = GetSlideId(i);
-                <div class="carousel-dot" id="dot-@slideId"></div>
-            }
-        </div>
-    </div>
-    <div class="c-side-right">
-        <div class="arrow">
-            <i class="icon-chevron-right text-5xl text-[#9294E1]"></i>
-        </div>
-    </div>
-</div>
-
-@code {
-    [CascadingParameter(Name = "CarouselLength")] public int CarouselLength { get; set; }
-    [CascadingParameter(Name = "CarouselOrder")] public int CarouselOrder { get; set; }
-
-    [Parameter] public RenderFragment? TopContent { get; set; }
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public int SlideOrder { get; set; }
-
-    private string GetCurrentSlideId()
-        => $"slide-{CarouselOrder}-{SlideOrder}";
-
-    private string GetSlideId(int id)
-        => $"slide-{CarouselOrder}-{id}";
-}
