--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingFeatures.razor
@@ -1,66 +1,0 @@
-<div class="c-row">
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-01.svg" alt="" class="md:!h-48">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Voice and text</span>
-            <span class="features-text">are blended together </span>
-            <span class="features-text">&ndash; everywhere!</span>
-        </TextContent>
-    </LandingFeaturesItem>
-
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-02.svg" alt="">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Markdown-style markup</span>
-            <span class="features-text">
-                **bold**, *italic*, `preformatted` text and ```code blocks```
-            </span>
-        </TextContent>
-    </LandingFeaturesItem>
-
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-03.svg" alt="">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Replies, @@mentions, and reactions</span>
-        </TextContent>
-    </LandingFeaturesItem>
-</div>
-
-<div class="c-row">
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-04.svg" alt="">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Group and peer-to-peer</span>
-            <span class="font-bold">chats</span>
-        </TextContent>
-    </LandingFeaturesItem>
-
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-07.png" alt="">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Unlimited image & file uploads</span>
-        </TextContent>
-    </LandingFeaturesItem>
-
-    <LandingFeaturesItem>
-        <ImageContent>
-            <img draggable="false" src="/dist/images/landing/dialog-06.svg" alt="">
-        </ImageContent>
-        <TextContent>
-            <span class="font-bold">Avatars</span>
-            <span class="features-text">
-                change name, picture, and bio on per-chat basis
-            </span>
-        </TextContent>
-    </LandingFeaturesItem>
-</div>
