--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingFeaturesItem.razor
@@ -1,13 +1,0 @@
-<div class="c-cell">
-    <div class="c-image">
-        @ImageContent
-    </div>
-    <div class="c-text">
-        @TextContent
-    </div>
-</div>
-
-@code {
-    [Parameter] public RenderFragment? ImageContent { get; set; }
-    [Parameter] public RenderFragment? TextContent { get; set; }
-}
