--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingItemDescription.razor
@@ -1,15 +1,0 @@
-<div class="c-item-description @Class">
-    @if (!Title.IsNullOrEmpty()) {
-        <h1 class="title-text @TitleClass">
-            @Title
-        </h1>
-    }
-    @ChildContent
-</div>
-
-@code {
-    [Parameter] public string Title { get; set; } = "";
-    [Parameter] public string TitleClass { get; set; } = "";
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-}
