--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingItemImage.razor
@@ -1,8 +1,0 @@
-<div class="c-item-image @Class">
-    @ChildContent
-</div>
-
-@code {
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-}
