--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingItemTitle.razor
@@ -1,8 +1,0 @@
-<h1 class="title-text @Class">
-    @ChildContent
-</h1>
-
-@code {
-    [Parameter] public string Class { get; set; } = "";
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-}
