+++ src/dotnet/UI.Blazor.App/Pages/Landing/LandingPage1.razor
@@ -1,7 +1,6 @@
 @namespace ActualChat.UI.Blazor
 @using ActualChat.UI.Blazor.App.Pages.Landing
 @using ActualChat.Hosting
-
 @{
     var isMaui = HostInfo.HostKind.IsMauiApp();
     var signinText = isMaui ? "Sign in" : "Sign-in to web app";
