+++ src/dotnet/UI.Blazor.App/Pages/Landing/LandingPage4.razor
@@ -1,6 +1,5 @@
 @namespace ActualChat.UI.Blazor
 @using ActualChat.UI.Blazor.App.Pages.Landing
-
 @{
     var isNarrow = ScreenSize.IsNarrow();
 }
