--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingTimeline.razor
@@ -1,87 +1,0 @@
-<div class="page page-timeline no-auto-scroll">
-    <div class="title">
-        <div>
-            <strong>Communication</strong>
-            enabled humans to evolve faster than any other form of life on Earth.
-        </div>
-    </div>
-    <div class="timeline">
-        <LandingTimelineItem TextTop="fire" TextBottom="400,000 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-fire.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem IsLeft="true" TextTop="talking languages" TextBottom="200,000 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-talking.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem TextTop="written languages" TextBottom="7,300 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-written.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem IsLeft="true" TextTop="wheel" TextBottom="5,300 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-wheel.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem TextTop="telegraph" TextBottom="180 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-telegraph.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem IsLeft="true" TextTop="telephone" TextBottom="146 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-telephone.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem TextTop="penicillin" TextBottom="95 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-penicillin.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem IsLeft="true" TextTop="digital computer" TextBottom="80 years ago" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-computer.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem TextTop="mobile phone" TextBottom="50 years ago" IconClass="pb-2" IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-mobile.svg" alt="" class="self-end">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-        <LandingTimelineItem
-            IsLeft="true"
-            TextBottom="10 years ago: dramatic progress in computer vision, speech recognition, and many other areas that were thought to be extremely hard for computers to tackle"
-            IsNarrow="@IsNarrow">
-            <ContentIcon>
-                <img draggable="false" src="/dist/images/landing/timeline-satellite.svg" alt="">
-            </ContentIcon>
-        </LandingTimelineItem>
-
-    </div>
-    <div class="flex overflow-hidden justify-center items-center @(IsNarrow ? "m-4 text-title-1" : "m-12 text-2xl")">
-        <div class="flex-y w-144 text-center">
-            <span>and here we are,</span>with text, audio, and video chats
-        </div>
-    </div>
-    <div class="flex overflow-hidden justify-center items-center @(IsNarrow ? "m-4 text-title-1" : "m-12 text-4xl")">
-        <div class="flex w-144 text-center">
-            But somehow we still have to choose whether to type or to talk!
-        </div>
-    </div>
-</div>
-
-@code {
-    [Parameter] public bool IsNarrow { get; set; }
-}
