--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingTimelineItem.razor
@@ -1,36 +1,0 @@
-<div class="timeline-item">
-    <TimelineItemLeft
-        IconClass="@IconClass"
-        IsLeft="@IsLeft"
-        IsNarrow="@IsNarrow"
-        TextTop="@TextTop"
-        TextBottom="@TextBottom">
-        <ChildContent>
-            @if (IsLeft) {
-                @ContentIcon
-            }
-        </ChildContent>
-    </TimelineItemLeft>
-    <TimelineItemRight
-        IconClass="@IconClass"
-        IsLeft="@IsLeft"
-        IsNarrow="@IsNarrow"
-        TextTop="@TextTop"
-        TextBottom="@TextBottom">
-        <ChildContent>
-            @if (!IsLeft) {
-                @ContentIcon
-            }
-        </ChildContent>
-    </TimelineItemRight>
-
-</div>
-
-@code {
-    [Parameter] public bool IsLeft { get; set; }
-    [Parameter] public RenderFragment? ContentIcon { get; set; }
-    [Parameter] public string TextTop { get; set; } = "";
-    [Parameter] public string TextBottom { get; set; } = "";
-    [Parameter] public string IconClass { get; set; } = "";
-    [Parameter] public bool IsNarrow { get; set; }
-}
