--- src/dotnet/UI.Blazor.App/Pages/Landing/LandingUpcomingItem.razor
@@ -1,13 +1,0 @@
-@using ActualChat.UI.Blazor.Services
-
-<div class="flex-x gap-x-2 @(ScreenSize.IsNarrow() ? "w-3/4 h-16" : "w-80 h-14")">
-    <div class="bg-primary h-full min-w-1 max-w-1 rounded-full"></div>
-    <div class="w-full self-center">
-        @Text
-    </div>
-</div>
-
-@code {
-    [Parameter] public string Text { get; set; } = "";
-    [CascadingParameter] public ScreenSize ScreenSize { get; set; }
-}
