--- src/dotnet/UI.Blazor.App/Pages/Landing/NewLanding/NewLandingHeader.razor
@@ -1,164 +1,0 @@
-@using ActualChat.UI.Blazor.App.Pages.Landing
-@using ActualChat.Hosting
-@inherits LandingApp
-@namespace ActualChat.UI.Blazor
-@{
-    var account = State.Value;
-    var isGuest = account is not {IsGuestOrNone: false };
-    var avatar = account.Avatar;
-}
-
-@* Header *@
-<div class="landing-header blur-bg">
-    @if (ScreenSize.IsNarrow()) {
-        <ButtonRound Class="menu-button flex-none" Click="@Click">
-            <i class="icon-menu text-2xl"></i>
-        </ButtonRound>
-        <div class="c-content">
-            <div class="light-header">
-                <div class="c-icon">
-                    <img draggable="false" src="/dist/images/landing/ac-icon-light.svg" alt="" class="light-icon">
-                </div>
-                <div class="c-title-wrapper">
-                    <div class="c-title">
-                        <img draggable="false" src="/dist/images/landing/ac-title-light.svg" alt="" class="light-title">
-                    </div>
-                    <div class="c-version">
-                        <img draggable="false" src="/dist/images/landing/ac-alpha-light.svg" alt="" class="light-alpha">
-                    </div>
-                </div>
-            </div>
-            <div class="dark-header">
-                <div class="c-icon">
-                    <img draggable="false" src="/dist/images/landing/ac-icon-dark.svg" alt="" class="dark-icon">
-                </div>
-                <div class="c-title-wrapper">
-                    <div class="c-title">
-                        <img draggable="false" src="/dist/images/landing/ac-title-dark.svg" alt="" class="dark-title">
-                    </div>
-                    <div class="c-version">
-                        <img draggable="false" src="/dist/images/landing/ac-alpha-dark.svg" alt="" class="dark-alpha">
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="c-right">
-            @if (isGuest) {
-                <ButtonRound Class="signin-button" Click="@(SignIn)">
-                    <i class="icon-log-in text-xl"></i>
-                </ButtonRound>
-            } else {
-                <button
-                    class="context-menu-btn md:mx-4"
-                    data-menu="@(MenuRef.New<LandingDocsMenu>().ToString())"
-                    data-menu-trigger="@MenuTrigger.Primary"
-                    data-menu-placement="@(FloatingPosition.BottomStart.ToPositionString())">
-                    <AvatarCircle
-                        Avatar="@avatar"
-                        Size="SquareSize.Size9"/>
-                </button>
-            }
-        </div>
-    } else {
-        <div class="c-content">
-            <div class="light-header">
-                <div class="c-icon">
-                    <img draggable="false" src="/dist/images/landing/ac-icon-light.svg" alt="" class="light-icon">
-                </div>
-                <div class="c-title-wrapper">
-                    <div class="c-title">
-                        <img draggable="false" src="/dist/images/landing/ac-title-light.svg" alt="" class="light-title">
-                    </div>
-                    <div class="c-version">
-                        <img draggable="false" src="/dist/images/landing/ac-alpha-light.svg" alt="" class="light-alpha">
-                    </div>
-                    <div class="c-btn-group">
-                        <div class="w-4"></div>
-                        <Button
-                            Click="@WhyUsClick"
-                            Class="btn-text btn-transparent why-us">Why us
-                        </Button>
-                        <Button
-                            Click="@PricingClick"
-                            Class="btn-text btn-transparent">Why now
-                        </Button>
-                        <Button
-                            Click="@(_ => History.NavigateTo(Links.Docs))"
-                            Class="btn-text btn-transparent">Docs
-                        </Button>
-                    </div>
-                </div>
-            </div>
-
-            <div class="dark-header">
-                <div class="c-icon">
-                    <img draggable="false" src="/dist/images/landing/ac-icon-dark.svg" alt="" class="dark-icon">
-                </div>
-                <div class="c-title-wrapper">
-                    <div class="c-title">
-                        <img draggable="false" src="/dist/images/landing/ac-title-dark.svg" alt="" class="dark-title">
-                    </div>
-                    <div class="c-version">
-                        <img draggable="false" src="/dist/images/landing/ac-alpha-dark.svg" alt="" class="dark-alpha">
-                    </div>
-                    <div class="c-btn-group">
-                        <div class="w-4"></div>
-                        <Button
-                            Click="@WhyUsClick"
-                            Class="btn-text btn-transparent why-us">Why us
-                        </Button>
-                        <Button
-                            Click="@PricingClick"
-                            Class="btn-text btn-transparent">Why now
-                        </Button>
-                        <Button
-                            Click="@(_ => History.NavigateTo(Links.Docs))"
-                            Class="btn-text btn-transparent">Docs
-                        </Button>
-                    </div>
-                </div>
-            </div>
-            <div class="c-btn-group">
-                @if (isGuest) {
-                    <button class="signin-button-group items-center lg-icon" @onclick="@SignIn">
-                        <span class="font-medium">Sign in</span>
-                        <i class="icon-monitor text-2xl font-thin"></i>
-                    </button>
-                    <DownloadButton
-                        Title="Get app"
-                        Click="DownloadClick"
-                        Class="download-app lg-icon"/>
-                    <button class="btn-download btn-to-main-page !hidden" @onclick="@BackToMainPageClick">
-                        Back to Main page
-                    </button>
-                } else {
-                    <div
-                        class="context-menu-btn"
-                        data-menu="@(MenuRef.New<LandingDocsMenu>().ToString())"
-                        data-menu-trigger="@MenuTrigger.Primary"
-                        data-menu-placement="@(FloatingPosition.BottomStart.ToPositionString())">
-                        <AvatarCircle
-                            Avatar="@avatar"
-                            Size="SquareSize.Size10"/>
-                    </div>
-                }
-            </div>
-        </div>
-    }
-</div>
-
-@code {
-    [Inject] private History History { get; init; } = null!;
-    [Inject] private ModalUI ModalUI { get; init; } = null!;
-    [Inject] private HostInfo HostInfo { get; init; } = null!;
-
-    [CascadingParameter] public ScreenSize ScreenSize { get; set; }
-    [Parameter] public EventCallback PricingClick { get; set; }
-    [Parameter] public EventCallback DownloadClick { get; set; }
-    [Parameter] public EventCallback BackToMainPageClick { get; set; }
-    [Parameter] public EventCallback WhyUsClick { get; set; }
-    [Parameter] public EventCallback Click { get; set; }
-
-    private Task SignIn()
-        => ModalUI.Show(new SignInModal.Model());
-}
