+++ src/dotnet/UI.Blazor.App/Pages/Landing/PremiumFeaturesModal.razor
@@ -44,7 +44,7 @@
                 </div>
                 <div class="c-right">
                     <div class="text-note">
-                        We'll release AI features one-by-one, so some of them will be available in Q4 2023.
+                        We'll release AI features one-by-one, so some of them will be available in Q2 2024.
                     </div>
                 </div>
             </div>
