--- src/dotnet/UI.Blazor.App/Pages/Landing/TimelineItemLeft.razor
@@ -1,46 +1,0 @@
-<div class="part-left flex-x w-1/2 border-r">
-    @if (IsLeft) {
-        @if (!IsNarrow) {
-            <div class="content flex-x w-full">
-                <div class="context-text flex-y w-full items-end justify-center">
-                    <div class="text-top">
-                        @TextTop
-                    </div>
-                    <div class="text-bottom text-end">
-                        @TextBottom
-                    </div>
-                </div>
-                <div class="content-icon @IconClass">
-                    @ChildContent
-                </div>
-            </div>
-        } else {
-            <div class="content flex-y gap-y-2 w-full">
-                <div class="content-icon flex items-center justify-center min-w-24 mx-2 @IconClass">
-                    @ChildContent
-                </div>
-                <div class="context-text flex-y w-full items-end justify-center">
-                    <div class="text-top">
-                        @TextTop
-                    </div>
-                    <div class="text-bottom text-end">
-                        @TextBottom
-                    </div>
-                </div>
-            </div>
-        }
-        <div class="flex-y min-w-4">
-            <div class="top h-1/2 border-b"></div>
-            <div class="bottom h-1/2 border-t"></div>
-        </div>
-    }
-</div>
-
-@code {
-    [Parameter] public bool IsLeft { get; set; }
-    [Parameter] public string TextTop { get; set; } = "";
-    [Parameter] public string TextBottom { get; set; } = "";
-    [Parameter] public string IconClass { get; set; } = "";
-    [Parameter] public bool IsNarrow { get; set; }
-    [Parameter] public RenderFragment? ChildContent { get; set; }
-}
