+++ src/dotnet/UI.Blazor.App/exports.ts
@@ -1,7 +1,7 @@
 // Please sort the imports alphabetically!
-export * from './Pages/Landing/landing';
 export * from './Pages/Landing/landing-left-menu';
 export * from './Pages/Landing/landing-video-modal';
 export * from './Pages/Landing/Docs/scrollToId';
-export * from './Pages/Landing/NewLanding/new-landing';
+export * from './Pages/Landing/Docs/setSettings';
+export * from './Pages/Landing/landing';
 export * from './Pages/Test/js-test-page';
