+++ src/dotnet/UI.Blazor.App/styles.css
@@ -1,3 +1,2 @@
 @import './Components/maui-index.css';
 @import './Pages/Landing/landing.css';
-@import './Pages/Landing/NewLanding/new-landing.css';
